# opt: convert parameterized lookup joins to placeholder scans