--- a/pkg/sql/opt/exec/execbuilder/testdata/update
+++ b/pkg/sql/opt/exec/execbuilder/testdata/update
@@ -1025,28 +1025,20 @@ quality of service: regular
 │
 └── • render
     │
-    └── • lookup join
-        │ sql nodes: <hidden>
-        │ kv nodes: <hidden>
-        │ regions: <hidden>
-        │ actual row count: 0
-        │ KV time: 0µs
-        │ KV rows decoded: 0
-        │ KV bytes read: 0 B
-        │ KV gRPC calls: 0
-        │ execution time: 0µs
-        │ estimated max memory allocated: 0 B
-        │ table: t137352@t137352_pkey
-        │ equality: ($1) = (k)
-        │ equality cols are key
-        │ locking strength: for update
-        │
-        └── • values
-              sql nodes: <hidden>
-              regions: <hidden>
-              actual row count: 1
-              execution time: 0µs
-              size: 1 column, 1 row
+    └── • scan
+          sql nodes: <hidden>
+          kv nodes: <hidden>
+          regions: <hidden>
+          actual row count: 0
+          KV time: 0µs
+          KV rows decoded: 0
+          KV bytes read: 0 B
+          KV gRPC calls: 0
+          estimated max memory allocated: 0 B
+          missing stats
+          table: t137352@t137352_pkey
+          spans: [/1 - /1]
+          locking strength: for update
 
 statement ok
 INSERT INTO t137352 VALUES (1, 2, 3);
@@ -1297,29 +1289,20 @@ quality of service: regular
                 │ equality: (k) = (a)
                 │ pred: a = 1
                 │
-                └── • lookup join
-                    │ sql nodes: <hidden>
-                    │ kv nodes: <hidden>
-                    │ regions: <hidden>
-                    │ actual row count: 1
-                    │ KV time: 0µs
-                    │ KV rows decoded: 1
-                    │ KV pairs read: 2
-                    │ KV bytes read: 8 B
-                    │ KV gRPC calls: 1
-                    │ execution time: 0µs
-                    │ estimated max memory allocated: 0 B
-                    │ estimated row count: 1
-                    │ table: t137352@t137352_pkey
-                    │ equality: ($1) = (k)
-                    │ equality cols are key
-                    │
-                    └── • values
-                          sql nodes: <hidden>
-                          regions: <hidden>
-                          actual row count: 1
-                          execution time: 0µs
-                          size: 1 column, 1 row
+                └── • scan
+                      sql nodes: <hidden>
+                      kv nodes: <hidden>
+                      regions: <hidden>
+                      actual row count: 1
+                      KV time: 0µs
+                      KV rows decoded: 1
+                      KV pairs read: 2
+                      KV bytes read: 8 B
+                      KV gRPC calls: 1
+                      estimated max memory allocated: 0 B
+                      estimated row count: 1
+                      table: t137352@t137352_pkey
+                      spans: [/1 - /1]
 
 statement ok
 DELETE FROM t137352 WHERE true;