… pkg/sql/opt/xform/testdata/coster/placeholder-scan
… pkg/sql/opt/xform/testdata/rules/generic