Parameterized lookup joins, which are used in generic query plans, are
now converted to placeholder scans in some cases. These placeholder
scans are more efficient because they cater to a narrower set of
features than general-purpose lookup joins and because they natively
support vectorization. See the new exploration rule for more details.

Release note: None