@mgartner 2025-06-17
Initial TPC-C micro-benchmark results show some promise:

```
name                    old time/op    new time/op    delta
TPCC/literal/new_order    19.4ms ± 5%    18.8ms ± 4%  -3.16%  (p=0.000 n=20+19)

name                    old alloc/op   new alloc/op   delta
TPCC/literal/new_order    3.64MB ± 5%    3.36MB ± 7%  -7.78%  (p=0.000 n=20+20)

name                    old allocs/op  new allocs/op  delta
TPCC/literal/new_order     24.2k ± 5%     22.6k ± 5%  -6.70%  (p=0.000 n=20+20)
```

Running all the txns now.