@mgartner 2025-06-20
I reran sysbench and measured no change:

```
name                                                      old queries/sec  new queries/sec  delta
SysbenchSettings/a=oltp_read_write/nodes=3/cpu=8/conc=64       31.6k ± 5%       31.6k ± 7%   ~     (p=0.631 n=36+39)

name                                                      old txns/sec     new txns/sec     delta
SysbenchSettings/a=oltp_read_write/nodes=3/cpu=8/conc=64       1.58k ± 5%       1.58k ± 7%   ~     (p=0.631 n=36+39)

name                                                      old ms/min       new ms/min       delta
SysbenchSettings/a=oltp_read_write/nodes=3/cpu=8/conc=64        11.7 ± 5%        11.6 ± 6%   ~     (p=0.672 n=38+39)

name                                                      old ms/avg       new ms/avg       delta
SysbenchSettings/a=oltp_read_write/nodes=3/cpu=8/conc=64        40.5 ± 5%        40.5 ± 7%   ~     (p=0.633 n=36+39)

name                                                      old ms/p95       new ms/p95       delta
SysbenchSettings/a=oltp_read_write/nodes=3/cpu=8/conc=64        62.7 ± 7%        64.4 ±16%   ~     (p=0.525 n=33+38)

name                                                      old ms/max       new ms/max       delta
SysbenchSettings/a=oltp_read_write/nodes=3/cpu=8/conc=64         269 ±60%         253 ±59%   ~     (p=0.387 n=39+39)
```