@mgartner 2025-06-20
I measured an almost 2% increase in tpcc throughput.

```
name                                                        old ops/sec  new ops/sec  delta
TPCC/tpcc-nowait/literal/w=1000/nodes=5/cpu=16/delivery        109 ± 3%     111 ± 5%  +1.89%  (p=0.002 n=30+30)
TPCC/tpcc-nowait/literal/w=1000/nodes=5/cpu=16/neworder      1.09k ± 3%   1.11k ± 5%  +1.88%  (p=0.002 n=30+30)
TPCC/tpcc-nowait/literal/w=1000/nodes=5/cpu=16/orderstatus     109 ± 3%     111 ± 5%  +1.89%  (p=0.002 n=30+30)
TPCC/tpcc-nowait/literal/w=1000/nodes=5/cpu=16/payment       1.09k ± 3%   1.11k ± 5%  +1.88%  (p=0.002 n=30+30)
TPCC/tpcc-nowait/literal/w=1000/nodes=5/cpu=16/stocklevel      109 ± 3%     111 ± 5%  +1.89%  (p=0.002 n=30+30)
TPCC/tpcc-nowait/literal/w=1000/nodes=5/cpu=16/total         2.50k ± 3%   2.55k ± 5%  +1.88%  (p=0.001 n=30+30)

name                                                        old ms/avg   new ms/avg   delta
TPCC/tpcc-nowait/literal/w=1000/nodes=5/cpu=16/delivery       73.1 ± 7%    71.4 ± 7%    ~     (p=0.063 n=30+29)
TPCC/tpcc-nowait/literal/w=1000/nodes=5/cpu=16/neworder        137 ± 4%     133 ± 4%  -2.29%  (p=0.000 n=30+30)
TPCC/tpcc-nowait/literal/w=1000/nodes=5/cpu=16/orderstatus    16.6 ± 6%    16.2 ± 6%  -2.39%  (p=0.001 n=30+29)
TPCC/tpcc-nowait/literal/w=1000/nodes=5/cpu=16/payment        36.1 ±10%    35.8 ±13%    ~     (p=0.817 n=30+27)
TPCC/tpcc-nowait/literal/w=1000/nodes=5/cpu=16/stocklevel     21.9 ± 4%    21.7 ± 3%    ~     (p=0.062 n=30+29)
TPCC/tpcc-nowait/literal/w=1000/nodes=5/cpu=16/total          79.9 ± 3%    78.4 ± 5%  -1.83%  (p=0.002 n=30+30)

name                                                        old ms/p50   new ms/p50   delta
TPCC/tpcc-nowait/literal/w=1000/nodes=5/cpu=16/delivery       69.3 ± 9%    68.0 ±11%    ~     (p=0.179 n=30+30)
TPCC/tpcc-nowait/literal/w=1000/nodes=5/cpu=16/neworder        131 ± 4%     128 ± 5%  -2.05%  (p=0.001 n=29+30)
TPCC/tpcc-nowait/literal/w=1000/nodes=5/cpu=16/orderstatus    9.78 ± 9%    9.63 ± 9%    ~     (p=0.175 n=30+29)
TPCC/tpcc-nowait/literal/w=1000/nodes=5/cpu=16/payment        30.9 ±16%    30.6 ±18%    ~     (p=1.000 n=30+28)
TPCC/tpcc-nowait/literal/w=1000/nodes=5/cpu=16/stocklevel     16.2 ± 4%    16.3 ± 4%    ~     (p=0.145 n=28+28)
TPCC/tpcc-nowait/literal/w=1000/nodes=5/cpu=16/total          60.8 ± 0%    60.0 ± 6%  -1.27%  (p=0.034 n=23+30)

name                                                        old ms/p95   new ms/p95   delta
TPCC/tpcc-nowait/literal/w=1000/nodes=5/cpu=16/delivery        125 ± 4%     122 ± 7%  -2.46%  (p=0.004 n=30+30)
TPCC/tpcc-nowait/literal/w=1000/nodes=5/cpu=16/neworder        238 ± 6%     232 ± 6%  -2.68%  (p=0.001 n=30+28)
TPCC/tpcc-nowait/literal/w=1000/nodes=5/cpu=16/orderstatus    52.9 ± 7%    51.5 ± 6%  -2.65%  (p=0.002 n=30+30)
TPCC/tpcc-nowait/literal/w=1000/nodes=5/cpu=16/payment        75.5 ± 0%    75.5 ± 0%    ~     (all equal)
TPCC/tpcc-nowait/literal/w=1000/nodes=5/cpu=16/stocklevel     57.6 ± 6%    56.3 ± 7%  -2.19%  (p=0.005 n=30+30)
TPCC/tpcc-nowait/literal/w=1000/nodes=5/cpu=16/total           207 ± 5%     202 ± 8%  -2.30%  (p=0.009 n=30+30)

name                                                        old ms/p99   new ms/p99   delta
TPCC/tpcc-nowait/literal/w=1000/nodes=5/cpu=16/delivery        159 ± 0%     157 ± 7%  -1.76%  (p=0.014 n=21+30)
TPCC/tpcc-nowait/literal/w=1000/nodes=5/cpu=16/neworder        290 ± 4%     285 ± 0%  -1.74%  (p=0.016 n=30+16)
TPCC/tpcc-nowait/literal/w=1000/nodes=5/cpu=16/orderstatus    82.5 ± 7%    79.7 ± 0%  -3.39%  (p=0.000 n=30+18)
TPCC/tpcc-nowait/literal/w=1000/nodes=5/cpu=16/payment         107 ± 6%     104 ± 5%  -2.27%  (p=0.001 n=30+28)
TPCC/tpcc-nowait/literal/w=1000/nodes=5/cpu=16/stocklevel     87.7 ± 5%    85.6 ± 8%  -2.40%  (p=0.005 n=30+30)
TPCC/tpcc-nowait/literal/w=1000/nodes=5/cpu=16/total           261 ± 4%     258 ±11%  -1.32%  (p=0.041 n=26+30)

name                                                        old ms/max   new ms/max   delta
TPCC/tpcc-nowait/literal/w=1000/nodes=5/cpu=16/delivery        271 ±11%     266 ±20%    ~     (p=0.204 n=30+28)
TPCC/tpcc-nowait/literal/w=1000/nodes=5/cpu=16/neworder        463 ±16%     460 ±17%    ~     (p=0.664 n=26+27)
TPCC/tpcc-nowait/literal/w=1000/nodes=5/cpu=16/orderstatus     194 ±13%     185 ±14%  -4.62%  (p=0.038 n=29+28)
TPCC/tpcc-nowait/literal/w=1000/nodes=5/cpu=16/payment         237 ±13%     238 ±20%    ~     (p=0.579 n=28+26)
TPCC/tpcc-nowait/literal/w=1000/nodes=5/cpu=16/stocklevel      203 ±16%     204 ±31%    ~     (p=0.323 n=27+28)
TPCC/tpcc-nowait/literal/w=1000/nodes=5/cpu=16/total           463 ±16%     460 ±17%    ~     (p=0.664 n=26+27)
```

----

To collect this I ran:

```
git checkout <before_sha>
./pkg/cmd/roachtest/roachstress.sh -c 30 tpcc-nowait/literal/w=1000/nodes=5/cpu=16

git checkout <after_sha>
./pkg/cmd/roachtest/roachstress.sh -c 30 tpcc-nowait/literal/w=1000/nodes=5/cpu=16

./tpcc_results.sh artifacts/<before_sha> > before.txt
./tpcc_results.sh artifacts/<after_sha> > after.txt
benchstat before.txt after.txt
```

Where `tpcc_results.sh` is this script: https://gist.github.com/mgartner/07171bf9d8b515ce992d1b20691c72c5